change: settle the three review notes on the shared detail page layout
The post tab bar's `margin-block-end: 0` lands on the same element as
`SectionTabs`' own rule, both single-class, so it only won because the
toolkit's stylesheet is emitted first. Doubling the selector makes it win
outright, and an import reorder can no longer move the header spacing.

`DetailPageTabs` and `DetailPageTabPanel` move into `detail-page-tabs.tsx`,
matching `report-page-tabs` next door. The layout no longer imports
`SectionTabPanel`, so `video-detail/stage.test.tsx` drops the mock it carried
for a page with no tabs — a mock the suite already passed without.

`DetailPageShell`'s test mock swallowed `...pageProps`, leaving the one thing
the component does beyond `clsx` unasserted. The mock now captures the rest
props and the test checks `visual` and `breadcrumbs` reach `Page`; removing the
spread fails it.

WOOA7S-1949
