+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/components/detail-page/detail-page-layout.tsx
@@ -1,13 +1,13 @@
 /**
  * External dependencies
  */
-import { SectionHeader, SectionTabPanel } from '@jetpack-premium-analytics/ui';
+import { SectionHeader } from '@jetpack-premium-analytics/ui';
 import clsx from 'clsx';
 /**
  * Internal dependencies
  */
 import styles from './detail-page-layout.module.scss';
-import type { SectionHeaderProps, SectionTabPanelProps } from '@jetpack-premium-analytics/ui';
+import type { SectionHeaderProps } from '@jetpack-premium-analytics/ui';
 import type { ReactNode } from 'react';
 
 /** What a detail page hands the layout's header, owned by the header's own props. */
@@ -59,36 +59,4 @@ export interface DetailPageSectionProps {
  */
 export function DetailPageSection( { children, className }: DetailPageSectionProps ) {
 	return <div className={ clsx( styles.section, className ) }>{ children }</div>;
-}
-
-/**
- * The tab root ships beside the panel so a detail route imports both through one
- * package specifier. Base UI's tabs context does not cross bundle copies of
- * `@wordpress/ui`, so the two must come from the same module instance.
- */
-export {
-	SectionTabs as DetailPageTabs,
-	type SectionTab as DetailPageTab,
-	type SectionTabsProps as DetailPageTabsProps,
-} from '@jetpack-premium-analytics/ui';
-
-export type DetailPageTabPanelProps< TabId extends string = string > =
-	SectionTabPanelProps< TabId >;
-
-/**
- * A `DetailPageSection` that is also one tab's panel, for a tabbed detail page.
- *
- * @param {DetailPageTabPanelProps} props - The component props.
- * @return The section, as a tab panel.
- */
-export function DetailPageTabPanel< TabId extends string = string >( {
-	value,
-	children,
-	className,
-}: DetailPageTabPanelProps< TabId > ) {
-	return (
-		<SectionTabPanel value={ value } className={ clsx( styles.section, className ) }>
-			{ children }
-		</SectionTabPanel>
-	);
 }
