+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/components/detail-page/index.ts
@@ -1,13 +1,15 @@
 export {
 	DetailPageLayout,
 	DetailPageSection,
-	DetailPageTabPanel,
-	DetailPageTabs,
 	type DetailPageHeaderSlots,
 	type DetailPageLayoutProps,
 	type DetailPageSectionProps,
+} from './detail-page-layout';
+export {
+	DetailPageTabPanel,
+	DetailPageTabs,
 	type DetailPageTab,
 	type DetailPageTabPanelProps,
 	type DetailPageTabsProps,
-} from './detail-page-layout';
+} from './detail-page-tabs';
 export { DetailPageShell, type DetailPageShellProps } from './detail-page-shell';
