+++ projects/packages/premium-analytics/routes/video-detail/stage.test.tsx
@@ -30,8 +30,6 @@ jest.mock( '@jetpack-premium-analytics/routing', () => ( {
 jest.mock( '@jetpack-premium-analytics/ui', () => ( {
 	DateFiltersPanel: () => <div>Date filters</div>,
 	SectionHeader: jest.requireActual( '../../packages/ui/src/section-header' ).SectionHeader,
-	// Unused here — the page has no tabs — but the shared detail-page layout imports it.
-	SectionTabPanel: ( { children }: { children: ReactNode } ) => <div>{ children }</div>,
 	StatsBreadcrumbs: jest.requireActual( '../../packages/ui/src/stats-breadcrumbs' )
 		.StatsBreadcrumbs,
 	StatsPageIcon: () => null,
